+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support for access to patches for dev/test instances in AutonomousDatabase
 - Support for metastore integration added to `dataflow`
+- Support for Cross-Region Copy of Volume Group Backups in `core` added
 
 ## 4.37.0 (July 28, 2021)